requirements.txt
# ── Core AutoGen v0.4 ───────────────────────────────────────────────────────
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0

# ── FastAPI & server ─────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ── HTTP clients (sync + async) ──────────────────────────────────────────────
aiohttp>=3.9.0          # async HTTP for Wikipedia / DDG Instant Answer / APIs
httpx>=0.25.0           # used internally by OpenAI-compatible client

# ── Web search ───────────────────────────────────────────────────────────────
ddgs>=9.11.4            # DuckDuckGo real web search (renamed from duckduckgo-search)

# ── Gmail API (install for production, mock mode works without) ──────────────
google-auth-oauthlib>=1.0.0     # OAuth 2.0 for Gmail
google-api-python-client>=2.0.0  # Gmail API client

# ── Environment & utilities ──────────────────────────────────────────────────
python-dotenv>=1.0.0
aiofiles>=23.0.0

# ── Telegram Bot API (aiohttp already installed — no extra SDK needed) ────────
# Add these to your .env:
#   TELEGRAM_BOT_TOKEN=<your_bot_token>       (from @BotFather)
#   TELEGRAM_DEFAULT_CHAT_ID=<your_chat_id>   (optional default target)

# ── Development / testing (optional) ─────────────────────────────────────────
pytest>=7.0.0
pytest-asyncio>=0.21.0
